Fixed path errors.
"." refers to the directory user is in, not the directory of where the
script is. We can get the script directory by using
"$(dirname "$(readlink -f "$0")")".

$0 is the name of the script.

readlink -f gives us the absolute path of the script which prevents
symbolic link errors.

dirname gives us the absolute path of the directory of the script.